+++ includes/class-functions.php
@@ -264,7 +264,15 @@ class Voxel_Toolkit_Functions {
                 'class' => 'Voxel_Toolkit_Suggest_Edits',
                 'file' => 'functions/class-suggest-edits.php',
                 'settings_callback' => array('Voxel_Toolkit_Suggest_Edits', 'render_settings'),
-            )
+            ),
+            'widget_css_injector' => array(
+                'name' => __('Widget CSS Class & ID', 'voxel-toolkit'),
+                'description' => __('Add CSS Class and ID fields to Voxel widgets (Navbar, User Bar, Advanced List) for custom styling of individual items.', 'voxel-toolkit'),
+                'class' => 'Voxel_Toolkit_Widget_CSS_Injector',
+                'file' => 'functions/class-widget-css-injector.php',
+                'settings_callback' => array('Voxel_Toolkit_Widget_CSS_Injector', 'render_settings'),
+                'always_enabled' => true,
+            ),
         );
 
         // Allow other plugins/themes to register functions